Makefile: remember $PWD even running under sudo
sudo, doas, and other privilege-elevation tools do not preserve the
value of the environment PWD. GNU make does not set $(PWD) itself, but
gets the value from the environment. Therefore, when running
"sudo make install" with GNU make, $(PWD) evaluates to nothing, which
causes problems in the sed call for the install target.
Each rule in the Makefile is interpreted by a shell, and new shells must set
the PWD environ variable (the working dir is inherited from wherever
make was called from). Therefore, we set $(PWD) so that it evaluates to ${PWD},
which will be dereferenced correctly by the shell.

However, with this new value of $(PWD), any rules which change directory
will have their new directory as PWD, and not the original directory
where make was called from. Therefore, for these rules, set a local
variable with the evaluation of ${PWD} before changing directory, and
then use that variable.
This fix is also made to the comp rules which use $(BASE) after changing
directory, in case the user has left the default of BASE = $(PWD).

These workarounds are done so that this Makefile is portable; it has
been tested with both BSD make and GNU make, and it remains POSIX.1-2017
compliant. While both BSD and GNU make do set a special variable with
the value of the current working directory - $(.CURDIR) and $(CURDIR),
respectively - these are not compatible. The unportable extension we
would need to simplify this would be

	PWD ?= $${PWD}

so that PWD is only reset if it's not set already - in our case, only
when running under sudo etc. We would also want some immediate-expansion
macros. However, those methods are not portable, but this patch is.